Add `feature:account:oauth` module and move ServerNameSuggester there
