+++ app/core/src/main/java/com/fsck/k9/KoinModule.kt
@@ -2,13 +2,13 @@ package com.fsck.k9
 
 import android.content.Context
 import app.k9mail.core.android.common.coreCommonAndroidModule
+import app.k9mail.feature.account.oauth.domain.usecase.ServerNameSuggester
 import com.fsck.k9.helper.Contacts
 import com.fsck.k9.helper.DefaultTrustedSocketFactory
 import com.fsck.k9.mail.ssl.LocalKeyStore
 import com.fsck.k9.mail.ssl.TrustManagerFactory
 import com.fsck.k9.mail.ssl.TrustedSocketFactory
 import com.fsck.k9.mailstore.LocalStoreProvider
-import com.fsck.k9.setup.ServerNameSuggester
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.GlobalScope
 import kotlinx.datetime.Clock
