+++ feature/account/oauth/src/test/java/app/k9mail/feature/account/oauth/domain/usecase/ServerNameSuggesterTest.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.setup;
+package app.k9mail.feature.account.oauth.domain.usecase;
 
 
 import app.k9mail.core.common.mail.Protocols;
@@ -10,7 +10,6 @@ import static org.junit.Assert.assertEquals;
 
 public class ServerNameSuggesterTest {
     private ServerNameSuggester serverNameSuggester;
-
 
     @Before
     public void setUp() throws Exception {
